refactor: 处理入口文件
