+++ .env.development
@@ -1,3 +1,3 @@
 REACT_APP_TEST1 = test-dev
-REACT_APP_BASE_API_URL = https://test.dancebox.cn/api/v2
+REACT_APP_BASE_API_URL = http://localhost:4000/api/v2
 # REACT_APP_TITLE = dev环境
