+++ src/Layout/AppHeader/components/ModifyPasswordModal.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren } from 'react';
 import { Modal, Form, Input } from 'antd';
 import { useRequest } from 'ahooks';
 import AccountService from 'src/services/system/account';
-import { ModifyPasswordDto } from 'src/pages/System/Account/types/modify.password.dto';
+import { ModifyPasswordDto } from 'src/pages/System/Account/types';
 
 const layout = {
   labelCol: { span: 6 },
