+++ src/pages/System/Access/types/index.ts
@@ -1,0 +1,3 @@
+export * from './access.req.dto';
+export * from './access.res.dto';
+export * from './query.access.dto';
