+++ src/pages/System/Account/types/index.ts
@@ -1,0 +1,6 @@
+export * from './account.dto';
+export * from './account.res.dto';
+export * from './account.role.res.dto';
+export * from './dispatch.account.role.dto';
+export * from './modify.password.dto';
+export * from './query.account.dto';
