+++ src/pages/System/Role/types/index.ts
@@ -1,0 +1,7 @@
+export * from './query.role.dto';
+export * from './role.api.res.dto';
+export * from './role.menu.api.req.dto';
+export * from './role.menu.auth.res.dto';
+export * from './role.menus.res.dto';
+export * from './role.req.dto';
+export * from './role.res.dto';
