+++ src/services/system/account/index.ts
@@ -1,8 +1,10 @@
 import BaseService from '../../base';
-import { AccountDto } from 'src/pages/System/Account/types/account.dto';
-import { ModifyPasswordDto } from 'src/pages/System/Account/types/modify.password.dto';
-import { QueryAccountDto } from 'src/pages/System/Account/types/query.account.dto';
-import { AccountListResDto } from 'src/pages/System/Account/types/account.res.dto';
+import {
+  AccountDto,
+  ModifyPasswordDto,
+  QueryAccountDto,
+  AccountListResDto,
+} from 'src/pages/System/Account/types';
 
 class AccountService extends BaseService {
   // 添加数据
